refactor(frontend): split library + playlists into api mixin modules
Extract the library and playlist domain methods from api.ts into
frontend/lib/api/library.ts and frontend/lib/api/playlists.ts as typed
mixins composed onto the ApiClient facade. Adds the ApiClientConstructor
mixin base type to core.ts. Pure structural move.

Refactor slice P34 (api-client-typing-split).
